requirements/python-requirements.txt
django
django-environ
django-extensions
gunicorn
psycopg2
pytz
raven